% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.parameter.R
\name{classify.parameter}
\alias{classify.parameter}
\title{\code{classify.parameter} shows parameters used for
missed deaths classification in \code{classify}}
\usage{
classify.parameter(xx)
}
\arguments{
\item{xx}{object  of type \code{list} generated by classify.}
}
\value{
data.frame with parameters of the
missed deaths classification by \code{classify()}
}
\description{
\code{classify.parameter} shows parameters used for
missed deaths classification in \code{classify}
}
\examples{
data(C61)
data(sp)
xx = classify(C61,sp,survYears = 30,fu_end=2016,md_cutoff = 4.0)
classify.Param(xx)

}
\seealso{
\code{\link{classify}}
 \code{\link{classify.plotSurv}}
 \code{\link{classify.regCoef}}
 \code{\link{classify.summary}}
}
