% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.plotSurv.R
\name{classify.plotSurv}
\alias{classify.plotSurv}
\title{\code{classify.plotSurv} plots relative survival
rates by follow-up time before and after
exclusion of missed deaths by \code{classify}}
\usage{
classify.plotSurv(xx)
}
\arguments{
\item{xx}{object  of type \code{list} generated by classify.}
}
\value{
ggplot object
}
\description{
\code{plotSurv} plots relative survival by follow-up
 time before and after exclusion of cases classified
 as 'missed' deaths in \code{classify}.
 Relative survival rates > 200\% are truncated.
}
\examples{
data(C61)
data(sp)
xx = classify(C61,sp,survYears = 30,fu_end=2016,md_cutoff = 4.0)
classify.plotSurv(xx)

}
\seealso{
\code{\link{classify}}
 \code{\link{classify.parameter}}
 \code{\link{classify.regCoef}}
 \code{\link{classify.summary}}
}
