% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify missed deaths in cancer registry data}
\usage{
classify(dat, sp, fu_end, survYears, md_cutoff = 3, periodWidth = 5)
}
\arguments{
\item{dat}{\code{data.frame} consisting of 11 variables:\cr
 \code{id} - (int) unique identifier for each data line,\cr
 \code{sex} -  (int) 1 = male, 2 = female, \cr
 \code{icd} - (chr) identifier for type of cancer (e.g.,
 according to ICD-10),\cr
 \code{stage} - (factor) cancer stage (e.g., "I", "II",
 "III", "IV", "Missing" for UICC stages),\cr
 \code{diagage} - (int) age at diagnosis in years,\cr
 \code{agr} - (factor) age group at diagnosis,\cr
 \code{dm} - (int) month of diagnosis, a number in \code{1:12},\cr
 \code{dy} - (int) calendar year of diagnosis (e.g., 2010),\cr
 \code{fm} - (int) month of follow-up end, a number in \code{1:12}, \cr
 \code{fy} - (int) calendar year of follow-up end (e.g., 2012),\cr
 \code{death} - (int) death at follow-up end (0 = alive, 1 = dead).\cr
To improve classification, it is recommended to use only one
combination of icd, sex, age group and stage.}

\item{sp}{\code{list} of two \code{data.frame}s named
\code{females} resp. \code{males}. Each \code{data.frame}
contains age- and calendar year-specific survival
probabilities of women resp. men in the underlying population.
Each column should represent one calendar year and should
consist of a numeric vector of length 100 providing  1-year
survival probabilities from ages 0 to 99 years. Each column
should be named as the year it represents (e.g., \code{"2004"}).
If the data in \code{dat} include one or more years for which
there is no corresponding column, the column corresponding to
the closest year prior to the missing year(s) is used.}

\item{fu_end}{Integer, indicating the last year of follow-up
(eg. 2018). The function assumes that follow-up is complete
through December 31st of this year. Patients with date of
deaths after this date are regarded as survivors.}

\item{survYears}{Integer, indicating maximum follow-up time
in years (e.g., \code{survYears = 30}). Cases diagnosed in the
time span \cr\code{fu_end - survYears - periodWidth + 1} to
\code{fu_end} \cr are included in relative survival calculations.}

\item{md_cutoff}{Numeric, (default \code{value md_cutoff =
3.0}) indicating the maximum accepted percentage of
classified missed deaths. When the percentage of missed deaths
classified by logistic regression exceeds \code{md_cutoff},
the scoring algorithm is applied. Scoring is also stopped
when \code{md_cutoff} is exceeded.}

\item{periodWidth}{Numeric, (default value \code{periodWidth = 5})
indicating the width of the analysis period for relative
survival estimation in years.\cr
For example, if \code{fu_end = 2016} and \code{periodWidth = 2}
the analysis period starts on 2015/01/01 and ends on 2016/12/31.}
}
\value{
\code{classify} returns a \code{list} containing
 four objects (\code{dat}, \code{survTab}, \code{Param} and
 \code{regCoef})\cr

\itemize{
   \item{\code{dat}:\cr \code{data.frame} with the input data and four}
additional columns:
 \itemize{
 \item{\code{fu.age}: Age at follow-up end in years}
 \item{\code{fu.time}: Follow-up time in years}
 \item{\code{missedDeath}: 0: not a missed deaths;
 1: a missed deaths}
 \item{\code{prob}: propability of beeing a
    missed deaths estimated by the logistic regression.
    If the scoring algorithm is appield \code{prob} is
    set to \code{NA}}
 }
   \item{\code{survTab}:\cr \code{data.frame}
   contaning estimates of relative survival before
   and after excluding 'missed deaths'. \code{survTab}
   is used by \code{plotSurv()} for plotting survival rates.}
   \item{\code{Param}:\cr \code{list} with input and
   runtime parameters. \code{Param} is displayed by \code{classify.Param()}
   }
   \item{\code{regCoef}:\cr \code{data.frame} with coefficients
   of the logistic regression. If \code{Pr(>|z|) > 0.05} for
   both fu.age and fu.time, the logistic regression
   was regarded as not significant and the scoring algorithm
   was applied for classification of missed deaths.}
}
}
\description{
Classify potential missed deaths by comparing
survival times of living and deceased persons.
}
\details{
This function compares length of follow-up time
and age at end of follow-up between living and
deceased persons in the input dataset (\code{dat}).
To increase the validity of these comparisons,
data should be stratified according to
prognostic factors. Currently, prognostic factors
reflect items typically available in epidemiological
cancer registry data: Sex, Age at diagnosis,
Diagnosis and Stage. The input dataset should include
only one combination of the values of these variables,
e.g., Women with stage I melanoma of the skin between
50 and 69 years old at diagnosis. However, to strengthen
the validity of the classification, we recommend a
stratum size of at least 500 persons.

Because the algorithm implemented here requires reliable
information on follow-up time for deceased persons, cases
with unknown date of diagnosis (e.g., Death Certificate-Only
[DCO] cases) should be excluded. Cases lost to follow-up must
be censored at the last date on which vital status was known.

The input dataset should include one entry per person.
Therefore, if the original data include multiple cases
per person, only the most recent case should be included.
If there are multiple, simultaneously diagnosed cases, the
most lethal diagnosis (determined by the average prognosis
in the respective population or from the literature) should
be included.

Within \code{classify}, relative survival is estimated by the
period approach (Gondos A, Brenner H, Holleczek B (2009)).Per
default we chose the last 5 years before follow-up end
\code{fu_end} as the analysis period. The width of the
analysis period can be changed by the parameter
\code{periodWidth}, and follow-up end is always set by
\code{fu_end}.


Whenever 1-year conditional relative survival is between 95\%
and 103\% in all follow-up years, we assume that survial in the
patient group is similar to survival in the general population.
In this case the search for missed deaths is not carried out.
}
\examples{
data(C61)
data(sp)
xx = classify(C61,sp,fu_end=2018,survYears=30,md_cutoff=4.0,periodWidth=4)
classify.summary(xx)
classify.plotSurv(xx)
classify.parameter(xx)
classify.regCoef(xx)
head(xx$dat)

}
\references{
\insertRef{Dahm2023}{misseddeaths}\cr\cr
 \insertRef{Gondos2009}{misseddeaths}
}
\seealso{
\code{\link{classify.parameter}}
 \code{\link{classify.plotSurv}}
 \code{\link{classify.regCoef}}
 \code{\link{classify.summary}}
}
