% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.regCoef.R
\name{classify.regCoef}
\alias{classify.regCoef}
\title{\code{classify.regCoef} shows coefficients and p-values of the logistic
regression used for the  classification
in \code{classify}}
\usage{
classify.regCoef(xx)
}
\arguments{
\item{xx}{object of type \code{list} generated by classify.}
}
\value{
data.frame with regression coefficients
}
\description{
\code{classify.regCoef} shows coefficients and p-values of the logistic
regression used for the  classification
in \code{classify}
}
\examples{
data(C61)
data(sp)
xx = classify(C61,sp,survYears = 30,fu_end=2016,md_cutoff = 4.0)
classify.regCoef(xx)

}
\seealso{
\code{\link{classify}}
 \code{\link{classify.parameter}}
 \code{\link{classify.plotSurv}}
 \code{\link{classify.summary}}
}
