% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.summary.R
\name{classify.summary}
\alias{classify.summary}
\title{Summary of missed deaths classification}
\usage{
classify.summary(xx)
}
\arguments{
\item{xx}{object  of type \code{list} generated by \code{classify()}.}
}
\value{
list object with the described values
}
\description{
\code{classify.summary()} calculates the number of patients,
survivors, deaths and missed deaths detected by
 \code{classify}
In addition, the percentage of classified missed
 deaths relative to the number of deaths is provided
}
\examples{
data(C61)
data(sp)
xx = classify(C61,sp,survYears = 30,fu_end=2016,md_cutoff = 4.0)
classify.Summary(xx)

}
\seealso{
\code{\link{classify}}
 \code{\link{classify.parameter}}
 \code{\link{classify.plotSurv}}
 \code{\link{classify.regCoef}}
}
